serializers.py
from rest_framework import serializers
from .models import Expense, Income

# Serializer for Expense model
class ExpenseSerializer(serializers.ModelSerializer):
    class Meta:
        model = Expense
        fields = '__all__'

# Serializer for Income model
class IncomeSerializer(serializers.ModelSerializer):
    class Meta:
        model = Income
        fields = '__all__'